# Binance: Improve wsHandleData performance