--- a/exchanges/binance/binance_test.go
+++ b/exchanges/binance/binance_test.go
@@ -1,16 +1,19 @@
 package binance
 
 import (
+	"bufio"
 	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
+	"os"
 	"reflect"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"github.com/thrasher-corp/gocryptotrader/common"
 	"github.com/thrasher-corp/gocryptotrader/common/key"
 	"github.com/thrasher-corp/gocryptotrader/core"
@@ -1962,6 +1965,24 @@ func TestGetDepositAddress(t *testing.T) {
 		t.Error("GetDepositAddress() error cannot be nil")
 	case mockTests && err != nil:
 		t.Error("Mock GetDepositAddress() error", err)
+	}
+}
+
+func BenchmarkWsHandleData(bb *testing.B) {
+	bb.ReportAllocs()
+	b.SetSaveTradeDataStatus(true)
+	fixture, err := os.Open("testdata/wsHandleData.json")
+	require.NoError(bb, err)
+	defer func() {
+		err = fixture.Close()
+		assert.NoError(bb, err)
+	}()
+	for i := 0; i < bb.N; i++ {
+		s := bufio.NewScanner(fixture)
+		for s.Scan() {
+			err = b.wsHandleData(s.Bytes())
+			assert.NoError(bb, err)
+		}
 	}
 }