… exchanges/binance/testdata/wsHandleData.json