# PR Description
Another PPROF PR 🎉  I noticed that Binance processing was taking awhile due to getting all enabled pairs on every update

| Before | After |
| ------ | ------ |
|<img height="500" alt="image" src="https://github.com/thrasher-corp/gocryptotrader/assets/9261323/1650b818-cf02-49fa-be1d-1c0f24d7f0ca"> | <img height="500" alt="image" src="https://github.com/thrasher-corp/gocryptotrader/assets/9261323/a6eee3c2-1753-48ee-9b90-23daabec0741"> |


- I've used @gbjk's way of inspecting json with `jsonparser` package to minimise the amount of `marshall` and `unmarshall`s. 
- I've used @shazbert's pair matcher technology 🤖 
- I've used @gbjk, @shazbert & @samuael's `types.Number`

Here's the benchmark data:
```
// old
BenchmarkWsHandleData-10    	     132	   9056070 ns/op	 2929851 B/op	    1098 allocs/op
// new
BenchmarkWsHandleData-10    	     763	   1565811 ns/op	  316703 B/op	     331 allocs/op
```

That's a 5.783x performance increase for 8 websocket update types. Now scale this across the 500 subscription limit on Binance over the span of hours 💥 

## To benchmark on `master`
- grab benchmark code `BenchmarkWsHandleData`
- copy `./exchanges/binance/testdata/wsHandleData.json`
- change to `master` branch
- plop it in
- run the benchmark

Fixes # (issue)
- [x] New feature (non-breaking change which adds functionality)

## How has this been tested

- [x] go test ./... -race
- [x] golangci-lint run
- [x] `BenchmarkWsHandleData`

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation and regenerated documentation via the documentation tool
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally and on Github Actions with my changes
- [x] Any dependent changes have been merged and published in downstream modules